requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-Migrate==3.1.0  # o la versión más reciente disponible
